pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openclaw-voice"
version = "0.1.0-dev"
description = "Quick-install browser voice layer for OpenClaw"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.0",
    "livekit-api",
    "livekit-plugins-elevenlabs",
    "livekit-plugins-openai",
    "livekit-plugins-silero",
    "faster-whisper",
    "flask>=3.0",
    "google-auth",
    "google-auth-oauthlib",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"
